requirements.txt
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
google-api-core==1.22.2
google-auth==1.21.3
google-cloud-language==1.3.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
numpy==1.19.3
oauthlib==3.1.0
packaging==20.4
pluggy==0.13.1
protobuf==3.13.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.1.2
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
toml==0.10.1
tqdm @ file:///tmp/build/80754af9/tqdm_1600709023549/work
tweepy==3.9.0
urllib3==1.25.10
zipp==3.4.0
